Rename package uauth to auth
